move servers away from utilities
