+++ erddapy/__init__.py
@@ -2,7 +2,8 @@
 Easier access to scientific data
 """
 
-from erddapy.erddapy import ERDDAP, servers
+from erddapy.erddapy import ERDDAP
+from erddapy.servers import servers
 
 
 __all__ = ["ERDDAP", "servers"]
