+++ tests/test_erddapy.py
@@ -6,6 +6,7 @@ from erddapy import ERDDAP
 
 
 @pytest.mark.web
+@pytest.mark.vcr()
 def test_erddap_requests_kwargs():
     """Test that an ERDDAP instance can have requests_kwargs attribute assigned
     and are passed to the underlying methods"""
@@ -14,10 +15,7 @@ def test_erddap_requests_kwargs():
     timeout_seconds = 1  # request timeout in seconds
     slowwly_milliseconds = (timeout_seconds + 1) * 1000
     slowwly_url = (
-        "http://slowwly.robertomurray.co.uk/delay/"
-        + str(slowwly_milliseconds)
-        + "/url/"
-        + base_url
+        f"https://flash.siwalik.in/delay/{slowwly_milliseconds}/url/{base_url}"
     )
 
     connection = ERDDAP(slowwly_url)
